requirements.txt
Pillow==10.1.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
reportlab==4.0.6
